+++ tests/Fixtures/StubFactory.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace SpipRemix\Component\Dbal\Test\Fixtures;
 
+use SpipRemix\Component\Dbal\Connection\ConnectionInterface;
 use SpipRemix\Component\Dbal\FactoryInterface;
 use SpipRemix\Component\Dbal\FieldInterface;
 use SpipRemix\Component\Dbal\SchemaInterface;
@@ -16,7 +17,12 @@ use SpipRemix\Component\Dbal\TableInterface;
  */
 class StubFactory implements FactoryInterface
 {
-    public static function createFromArray(array $definitions): SchemaInterface
+    public function createConnectionFromArray(array $connection): ConnectionInterface
+    {
+        return new MockConnection($connection);
+    }
+
+    public function createSchemaFromArray(array $definitions): SchemaInterface
     {
         return new StubSchema;
     }