+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Factory de schéma
 - Convertisseur des globales SPIP<=4
 - ConnectionTrait, détecteur d'exensions PHP
-- ConnectionInterface
+- ConnectionInterface, File, Socket, TCP
 
 ### Removed